proph_couture_project/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
# mysqlclient==2.2.0
Pillow>=10.2.0
python-decouple==3.8
django-filter==23.3
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
pyjwt==2.8.0
setuptools==68.2.2  # Important !
wheel==0.41.2
psycopg2-binary>=2.9.9
django-phonenumber-field[phonenumbers]==7.1.0
drf-extra-fields>=0.15.0

# Social Auth
django-allauth==0.57.0
dj-rest-auth==5.0.2
requests==2.31.0
pyotp
qrcode
gunicorn
whitenoise
dj-database-url
